Code cleanup and modernization w/ newer paradigms
